backend/requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
mistralai
langchain_openai
python-dotenv
python-multipart
chromadb
ipykernel
python-jose[cryptography]
passlib[bcrypt]
psycopg2
httpx
sqlalchemy
databases
asyncpg
pydantic[email]
pinecone-client
langchain_pinecone
langchain_mistralai
ipywidgets
pypdf
html2text
ragas
llama_index
xmltodict
langchain_experimental
langchain_community